+++ app/models/api/academies_api/trust.rb
@@ -18,16 +18,16 @@ class Api::AcademiesApi::Trust < Api::BaseApiModel
 
   def self.attribute_map
     {
-      ukprn: "giasData.ukprn",
-      group_identifier: "giasData.groupId",
-      original_name: "giasData.groupName",
-      companies_house_number: "giasData.companiesHouseNumber",
-      address_street: "giasData.groupContactAddress.street",
-      address_locality: "giasData.groupContactAddress.locality",
-      address_additional: "giasData.groupContactAddress.additionalLine",
-      address_town: "giasData.groupContactAddress.town",
-      address_county: "giasData.groupContactAddress.country",
-      address_postcode: "giasData.groupContactAddress.postcode"
+      ukprn: "ukprn",
+      group_identifier: "referenceNumber",
+      original_name: "name",
+      companies_house_number: "companiesHouseNumber",
+      address_street: "address.street",
+      address_locality: "address.locality",
+      address_additional: "address.additionalLine",
+      address_town: "address.town",
+      address_county: "address.country",
+      address_postcode: "address.postcode"
     }
   end